Testscript

+++ DroneControlConsole/src/fieldtests/FieldTestScript.java
@@ -89,7 +89,7 @@ public abstract class FieldTestScript {
 		deploy(m,ips);
 	}
 	
-	protected boolean arrivedAtWaypoint(ArrayList<String> ips, Waypoint wp, int distanceThresholdInMeters) {
+	protected boolean arrivedAtWaypoint(ArrayList<String> ips, Waypoint wp, double distanceThresholdInMeters) {
 		
 		DronesSet set = console.getDronesSet();
 		
@@ -104,7 +104,7 @@ public abstract class FieldTestScript {
 		return true;
 	}
 	
-	protected GeoFence defineGeoFence(LatLon center, int width, int height) {
+	protected GeoFence defineGeoFence(LatLon center, double width, double height) {
 		
 		GeoFence fence = new GeoFence("fence");
 		
@@ -138,7 +138,7 @@ public abstract class FieldTestScript {
 		mapPanel.replaceEntities(newEntities);
 	}
 	
-	protected ArrayList<Waypoint> generateWaypointsInGeoFence(GeoFence fence, int number, int minDistance, int maxDistance, long randomSeed) {
+	protected ArrayList<Waypoint> generateWaypointsInGeoFence(GeoFence fence, int number, double minDistance, double maxDistance, long randomSeed) {
 		ArrayList<Waypoint> wps = new ArrayList<Waypoint>();
 		
 		Random r = new Random(randomSeed);
